requirements.txt
Pillow
cv2
pydub
simpleaudio
d3dshot
valclient